requirements.txt
rich==13.9.4
questionary==2.0.1
prompt_toolkit==3.0.48
mysql-connector-python==9.0.0
keyring==25.3.0
passlib==1.7.4
bcrypt==4.2.0
cffi==1.17.1
